+++ src/main/java/plugily/projects/murdermystery/arena/managers/ScoreboardManager.java
@@ -63,6 +63,7 @@ public class ScoreboardManager {
    * @see User
    */
   public void createScoreboard(User user) {
+    plugin.getRewardsHandler().performReward(user.getPlayer(), Reward.RewardType.SCOREBOARD_ADD);
     final String boardTitle = plugin.getChatManager().colorMessage("Scoreboard.Title");
     Scoreboard scoreboard = ScoreboardLib.createScoreboard(user.getPlayer()).setHandler(new ScoreboardHandler() {
       @Override