+++ src/main/java/plugily/projects/murdermystery/arena/role/Role.java
@@ -77,7 +77,8 @@ public enum Role {
       case DETECTIVE:
         return arena.isCharacterSet(Arena.CharacterType.DETECTIVE) && arena.getDetectiveList().contains(player);
       case FAKE_DETECTIVE:
-        return arena.isCharacterSet(Arena.CharacterType.FAKE_DETECTIVE) && player.equals(arena.getCharacter(Arena.CharacterType.FAKE_DETECTIVE));
+        Player character = arena.getCharacter(Arena.CharacterType.FAKE_DETECTIVE);
+        return character != null && player.equals(character);
       case MURDERER:
         return arena.isCharacterSet(Arena.CharacterType.MURDERER) && arena.getMurdererList().contains(player);
       case ANY_DETECTIVE: