+++ src/main/java/plugily/projects/murdermystery/commands/arguments/game/JoinArguments.java
@@ -30,7 +30,10 @@ import plugily.projects.murdermystery.commands.arguments.data.CommandArgument;
 import plugily.projects.murdermystery.handlers.ChatManager;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
@@ -49,6 +52,22 @@ public class JoinArguments {
           sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.Type-Arena-Name"));
           return;
         }
+        if(args[1].equalsIgnoreCase("maxplayers") && ArenaRegistry.getArena("maxplayers") == null) {
+          Map<Arena, Integer> arenas = new HashMap<>();
+          for(Arena arena : ArenaRegistry.getArenas()) {
+            arenas.put(arena, arena.getPlayers().size());
+          }
+          if(ArenaRegistry.getArenaPlayersOnline() == 0) {
+            ArenaManager.joinAttempt((Player) sender, ArenaRegistry.getArenas().get(ThreadLocalRandom.current().nextInt(ArenaRegistry.getArenas().size())));
+            return;
+          }
+          Stream<Map.Entry<Arena, Integer>> sorted = arenas.entrySet().stream().sorted(Map.Entry.comparingByValue());
+          if(sorted.findFirst().isPresent()) {
+            Arena arena = sorted.findFirst().get().getKey();
+            ArenaManager.joinAttempt((Player) sender, arena);
+            return;
+          }
+        }
         for(Arena arena : ArenaRegistry.getArenas()) {
           if(args[1].equalsIgnoreCase(arena.getId())) {
             ArenaManager.joinAttempt((Player) sender, arena);
@@ -64,29 +83,20 @@ public class JoinArguments {
       registry.mapArgument("murdermystery", new CommandArgument("randomjoin", "", CommandArgument.ExecutorType.PLAYER) {
         @Override
         public void execute(CommandSender sender, String[] args) {
-          //first random get method
-          Map<Arena, Integer> arenas = new HashMap<>();
-          for(Arena arena : ArenaRegistry.getArenas()) {
-            if(arena.getArenaState() == ArenaState.STARTING && arena.getPlayers().size() < arena.getMaximumPlayers()) {
-              arenas.put(arena, arena.getPlayers().size());
-            }
+          //check starting arenas -> random
+          List<Arena> arenas = ArenaRegistry.getArenas().stream().filter(arena -> arena.getArenaState() == ArenaState.STARTING && arena.getPlayers().size() < arena.getMaximumPlayers()).collect(Collectors.toList());
+          if(!arenas.isEmpty()) {
+            Arena arena = arenas.get(ThreadLocalRandom.current().nextInt(arenas.size()));
+            ArenaManager.joinAttempt((Player) sender, arena);
+            return;
           }
-          if(arenas.size() > 0) {
-            Stream<Map.Entry<Arena, Integer>> sorted = arenas.entrySet().stream().sorted(Map.Entry.comparingByValue());
-            Arena arena = sorted.findFirst().get().getKey();
-            if(arena != null) {
-              ArenaManager.joinAttempt((Player) sender, arena);
-              return;
-            }
-          }
-
-          //fallback safe method
-          for(Arena arena : ArenaRegistry.getArenas()) {
-            if((arena.getArenaState() == ArenaState.WAITING_FOR_PLAYERS || arena.getArenaState() == ArenaState.STARTING)
-              && arena.getPlayers().size() < arena.getMaximumPlayers()) {
-              ArenaManager.joinAttempt((Player) sender, arena);
-              return;
-            }
+          //check waiting arenas -> random
+          arenas = ArenaRegistry.getArenas().stream().filter(arena -> (arena.getArenaState() == ArenaState.WAITING_FOR_PLAYERS || arena.getArenaState() == ArenaState.STARTING)
+              && arena.getPlayers().size() < arena.getMaximumPlayers()).collect(Collectors.toList());
+          if(!arenas.isEmpty()) {
+            Arena arena = arenas.get(ThreadLocalRandom.current().nextInt(arenas.size()));
+            ArenaManager.joinAttempt((Player) sender, arena);
+            return;
           }
           sender.sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Commands.No-Free-Arenas"));
         }